+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - `pressure_drops` file pattern extension changed to `dat` (@mkasztelnik)
 - JWT expiration time is now the same as for other envs (@mkasztelnik)
 - JWT expiration time can be configured using ENV variable (@mkasztelnik)
+- `stage_in` adds commented line when file cannot be found (@mkasztelnik)
 
 ### Deprecated